Debugger: Workaround for branch symbol name truncation
Adds a workaround for an issue where including a space or any comparison sign before a branch's symbol name would display an ellipses instead of the symbol name.

This commit replaces "--> " with 4 underscores to avoid this behavior.

Unsure how to properly fix this. Worth noting that any symbol names that contain a space or a comparison sign will break truncation as well.